+++ FWI/Python/daily_summaries.py
@@ -133,7 +133,7 @@ def generate_daily_summaries(hourly_FWI, reset_hr = 5,
       else:
         peak_time = by_date["isi_smooth"].idxmax()
       
-      # find the rest of the values at peak
+      # append date
       results["id"].append(stn)
       results["yr"].append(by_date.at[0, "yr"])
       results["mon"].append(by_date.at[0, "mon"])
@@ -142,12 +142,30 @@ def generate_daily_summaries(hourly_FWI, reset_hr = 5,
       # format sunrise and sunset as hh:mm from decimal hours
       sr = by_date.at[peak_time, "sunrise"]
       ss = by_date.at[peak_time, "sunset"]
-      results["sunrise"].append("{:02d}:{:02d}".format(int(sr), int(60 * (sr - int(sr)))))
-      results["sunset"].append("{:02d}:{:02d}".format(int(ss), int(60 * (ss - int(ss)))))
+      results["sunrise"].append("{:02d}:{:02d}".format(int(sr),
+        int(60 * (sr - int(sr)))))
+      results["sunset"].append("{:02d}:{:02d}".format(int(ss),
+        int(60 * (ss - int(ss)))))
       
+      # append hour of peak burn and active burning duration
       results["peak_hr"].append(by_date.at[peak_time, "hr"])
-      results["duration"].append(sum(by_date["isi_smooth"] > Spread_Threshold_ISI))
+      # calculate duration of active burning window
+      if any(by_date["isi_smooth"] >= Spread_Threshold_ISI):
+        # find first and last hours of active burning
+        active_burning = by_date[by_date["isi_smooth"] >= Spread_Threshold_ISI]
+        t_ab0 = datetime.datetime(active_burning.iloc[0].yr,
+          active_burning.iloc[0].mon,
+          active_burning.iloc[0].day,
+          active_burning.iloc[0].hr)
+        t_ab1 = datetime.datetime(active_burning.iloc[-1].yr,
+          active_burning.iloc[-1].mon,
+          active_burning.iloc[-1].day,
+          active_burning.iloc[-1].hr)
+        results["duration"].append((t_ab1 - t_ab0).seconds // 3600 + 1)
+      else:
+        results["duration"].append(0)
 
+      # append outputs at peak burn
       results["ffmc"].append(by_date.at[peak_time, "ffmc"])
       results["dmc"].append(by_date.at[peak_time, "dmc"])
       results["dc"].append(by_date.at[peak_time, "dc"])